Symbolic Link Command Fix (#645)
All relative symbolic link commands are currently broken and receive the response "bash: command not found: x" in the honeypot shell.  The path is not being traversed correctly and fs.HoneyPotFilesystem.getfile() appears to be the source of problem.  Removing the cwd join command resolves the problem but I am not aware of other potential side effects.