Docker output (#869)
* Full Docker support for output plugins

cowrie makes more fun if there are also all the output plugins available
inside the docker image. And now they are!

* Fixing wrong requirements: snappy vs python-snappy

snappy: “SnapPy is a package for studying the topology and geometry of 3-manifolds, with a focus on hyperbolic structures. It is based on the SnapPea kernel written by Jeff Weeks.”

python-snappy: “Python bindings for the snappy compression library from Google.”

We want python-snappy. ;)

* MAINTAINER argument deprecated in Docker

The MAINTAINER keyword is deprecated in Docker and is replaced with the
new syntax.
Also it is now attached to the right container and not to one we
actually throw away.